test(LottosTest): 테스트 케이스 위치 변경
